refactor(earn): break pool info screen into multiple files
