+++ src/earn/poolInfoScreen/TokenIcons.tsx
@@ -1,0 +1,37 @@
+import React from 'react'
+import { StyleSheet, View } from 'react-native'
+import TokenIcon, { IconSize } from 'src/components/TokenIcon'
+import { Spacing } from 'src/styles/styles'
+import { TokenBalance } from 'src/tokens/slice'
+
+export default function TokenIcons({
+  tokensInfo,
+  size = IconSize.MEDIUM,
+  showNetworkIcon = true,
+}: {
+  tokensInfo: TokenBalance[]
+  size?: IconSize
+  showNetworkIcon?: boolean
+}) {
+  return (
+    <View style={styles.container}>
+      {tokensInfo.map((token, index) => (
+        <TokenIcon
+          key={token.tokenId}
+          token={token}
+          viewStyle={!!index && { marginLeft: -Spacing.Regular16, zIndex: -index }}
+          size={size}
+          showNetworkIcon={showNetworkIcon}
+        />
+      ))}
+    </View>
+  )
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+})
